Rename sub-package to `pytorch` to avoid conflict with `torch` package
