+++ src/poprox_recommender/diversifiers/pfar.py
@@ -3,9 +3,9 @@ import math
 import torch as th
 
 from poprox_concepts import Article, ArticleSet, InterestProfile
+from poprox_recommender.pytorch.decorators import torch_inference
 from poprox_recommender.rankers import Ranker
 from poprox_recommender.topics import GENERAL_TOPICS, extract_general_topics, normalized_topic_count
-from poprox_recommender.torch.decorators import torch_inference
 
 
 class PFARDiversifier(Ranker):
