+++ src/poprox_recommender/embedders/article.py
@@ -7,8 +7,8 @@ import torch as th
 from transformers import PreTrainedTokenizer
 
 from poprox_concepts import ArticleSet
-from poprox_recommender.torch.datachecks import assert_tensor_size
-from poprox_recommender.torch.decorators import torch_inference
+from poprox_recommender.pytorch.datachecks import assert_tensor_size
+from poprox_recommender.pytorch.decorators import torch_inference
 
 logger = logging.getLogger(__name__)
 TITLE_LENGTH_LIMIT = 30
